added protection with jwt in task routes

+++ backend/controllers/taskController.js
@@ -1,9 +1,10 @@
 const asyncHandler = require('express-async-handler');
 const Task = require('../models/taskModel');
+const User = require('../models/userModel');
 
 const getTasks = asyncHandler(async (req, res) => {
     //res.status(200).json({message: 'Get all tasks'});
-    const tasks = await Task.find();
+    const tasks = await Task.find({ user: req.user.id });
     res.status(200).json(tasks);
 });
 
@@ -14,16 +15,26 @@ const setTask = asyncHandler(async (req, res) => {
         throw new Error('Please enter a task');
     }
     //res.status(200).json({message: 'Create Task'});
-    const task = await Task.create({ text: req.body.text })
+    const task = await Task.create({ text: req.body.text, user: req.user.id })
 
     res.status(200).json(task);
 });
 
 const updateTask = asyncHandler(async (req, res) => {
-    const task = Task.findById(req.params.id);
+    const task = await Task.findById(req.params.id);
     if(!task) {
         res.status(400);
         throw new Error('Task not found');
+    }
+    const user = await User.findById(req.user.id);
+    if(!user) {
+        res.status(401);
+        throw new Error('No such user found.');
+    }
+    //console.log(task.id);
+    if(task.user.toString() !== user.id) {
+        res.status(401);
+        throw new Error('User is not authorized to update.')
     }
     const updatedTask = await
         Task.findByIdAndUpdate(req.params.id, req.body, {
@@ -38,6 +49,15 @@ const deleteTask = asyncHandler(async (req, res) => {
     if(!task) {
         res.status(400);
         throw new Error('task not found');
+    }
+    const user = await User.findById(req.user.id);
+    if(!user) {
+        res.status(401);
+        throw new Error('No such user found.');
+    }
+    if(task.user.toString() !== user.id) {
+        res.status(401);
+        throw new Error('User is not authorized to delete.')
     }
     await Task.findByIdAndDelete(req.params.id);
     res.status(200).json({id: req.params.id});
